handle reseting queue items on selecting some new songs

+++ app/src/main/kotlin/com/simplemobiletools/musicplayer/models/QueueItem.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "queue_items", indices = [(Index(value = ["id"], unique = true))])
 data class QueueItem(
     @PrimaryKey(autoGenerate = true) var id: Int,
-    @ColumnInfo(name = "track_id") var trackId: Int,
+    @ColumnInfo(name = "track_id") var trackId: Long,
     @ColumnInfo(name = "track_order") var trackOrder: Int,
     @ColumnInfo(name = "is_playing") var isPlaying: Boolean,
     @ColumnInfo(name = "last_position") var lastPosition: Int,
